+++ src/Mutex.php
@@ -2,12 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Utopia\Lock\Lock;
+namespace Utopia\Lock;
 
 use Swoole\Coroutine;
 use Swoole\Coroutine\Channel;
 use Utopia\Lock\Exception\Contention;
-use Utopia\Lock\Lock;
 
 final class Mutex implements Lock
 {