+++ tests/MutexTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Utopia\Tests\Lock;
+namespace Utopia\Lock\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Swoole\Coroutine;
 use Swoole\Coroutine\System;
 use Utopia\Lock\Exception\Contention;
-use Utopia\Lock\Lock\Mutex;
+use Utopia\Lock\Mutex;
 
 use function Swoole\Coroutine\run;